// src/app/interface/employee-interface.ts
export interface EmployeeInterface {
  id: number;
  firstName: string;
  lastName: string;
  identificationDocument: string;
  expeditionPlace: string;
  residenceAdress: string;
  city: string;
  departament: string;
  country: string;
  email: string;
  education: string;
  academicTitle: string;
  archivo: string;
}